Many many many fixes, many remain....
Fixed checking that expected all things from front to be string and failed for any data if it was not string even tho it was valid.

Fixed global achievements stats to actually work, before there was not even a route....

Added a subscriber that shoudl broadcast final game results to notifications tab/dropdown, did not test if it works.